♻️(api) split api code by endpoints
The api.py file had become too long so it is better to split it
to several files.
