+++ src/magnify/apps/core/api/__init__.py
@@ -1,0 +1,36 @@
+"""Magnify core API endpoints"""
+from django.core.exceptions import ValidationError
+
+from rest_framework import exceptions as drf_exceptions
+from rest_framework import views as drf_views
+
+from .groups import GroupViewSet
+from .rooms import RoomViewSet
+from .users import UserViewSet
+
+# Necessary precision to comply with PEP8
+__all__ = [
+    "GroupViewSet",
+    "RoomViewSet",
+    "UserViewSet",
+]
+
+
+def exception_handler(exc, context):
+    """Handle Django ValidationError as an accepted exception.
+
+    For the parameters, see ``exception_handler``
+    This code comes from twidi's gist:
+    https://gist.github.com/twidi/9d55486c36b6a51bdcb05ce3a763e79f
+    """
+    if isinstance(exc, ValidationError):
+        if hasattr(exc, "message_dict"):
+            detail = exc.message_dict
+        elif hasattr(exc, "message"):
+            detail = exc.message
+        elif hasattr(exc, "messages"):
+            detail = exc.messages
+
+        exc = drf_exceptions.ValidationError(detail=detail)
+
+    return drf_views.exception_handler(exc, context)
